Threadsafe db

+++ includes/wireddb.py
@@ -15,6 +15,7 @@ class wiredDB():
 
     ## DB HANDLING ##
     def openDB(self):
+        self.lock.acquire()
         self.conn = sqlite3.connect(self.config["dbFile"], check_same_thread=True)
         self.pointer = self.conn.cursor()
         self.conn.text_factory = str
@@ -26,6 +27,7 @@ class wiredDB():
             return false
         self.conn.close()
         self.dbIsOpen = 0
+        self.lock.release()
         return 1
 
     ## User Managment ##
@@ -41,14 +43,13 @@ class wiredDB():
 
     def updateElement(self, data, type):
         if not self.openDB():
+            self.lock.release()
             return 0
-        sql = "UPDATE wiredUsers SET password='" + str(data[1]) + "', groupname='" + str(data[2]) + "', privsvs='" +\
+        sql = "UPDATE wiredUsers SET password='" + str(data[1]) + "', groupname='" + str(data[2]) + "', privs='" +\
             str(data[3]) + "' WHERE name='" + str(data[0]) + "' AND type='" + str(type) + "';"
         result = self.pointer.execute(sql)
-        self.lock.acquire()
         self.conn.commit()
         self.closeDB()
-        self.lock.release()
         if not result:
             self.logger.error("db failed to update a row")
             return 0
@@ -56,13 +57,12 @@ class wiredDB():
 
     def deleteElement(self, username, type):
         if not self.openDB():
+            self.lock.release()
             return 0
         sql = "DELETE FROM wiredUsers WHERE name='" + str(username) + "' AND type='" + str(type) + "';"
         result = self.pointer.execute(sql)
-        self.lock.acquire()
         self.conn.commit()
         self.closeDB()
-        self.lock.release()
         if not result:
             self.logger.error("db failed to delete a row")
             return 0
@@ -70,6 +70,7 @@ class wiredDB():
 
     def loadData(self, type):
         if not self.openDB():
+            self.lock.release()
             return 0
         # since there is no check if a table exists try to create basic userdb everytime we open the db
         self.pointer.execute('CREATE TABLE IF NOT EXISTS wiredUsers (name TEXT UNIQUE, password TEXT, groupname \
@@ -90,53 +91,45 @@ class wiredDB():
                                   chr(28) + '0' + chr(28) + '0' + chr(28) + '0'])
         self.conn.commit()
         data = 0
-        self.lock.acquire()
         self.pointer.execute("SELECT * FROM wiredUsers WHERE type=?;", [type])
         data = self.pointer.fetchall()
         if len(data) == 0:
             self.logger.info("userdb returned empty result for type %s", type)
             data = []
         self.closeDB()
-        self.lock.release()
         return data
 
     def saveUser(self, user):
         if not self.openDB():
+            self.lock.release()
             return 0
-        self.lock.acquire()
         self.pointer.execute("SELECT * FROM wiredUsers WHERE name = '" + str(user[0]) +
                              "' AND type ='" + str(user[3]) + "';")
         existing = self.pointer.fetchall()
-        self.lock.release()
         if existing:
             # user already exists!
             self.closeDB()
             return 0
-        self.lock.acquire()
         self.pointer.execute("INSERT INTO wiredUsers VALUES ('" + str(user[0]) + "','" + str(user[1]) + "', '" +
                              str(user[2]) + "', '" + str(user[3]) + "', '" + str(user[4]) + "');")
         self.conn.commit()
-        self.lock.release()
         self.closeDB()
         return 1
 
     ### News ###
     def loadNews(self):
         if not self.openDB():
+            self.lock.release()
             return 0
         sql = 'create table if not exists wiredNews (nick TEXT, date FLOAT, news TEXT)'
-        self.lock.acquire()
         self.pointer.execute(sql)
         self.conn.commit()
-        self.lock.release()
         data = 0
-        self.lock.acquire()
         self.pointer.execute("SELECT * FROM wiredNews")
         data = self.pointer.fetchall()
         if not data:
             self.logger.error("db returned empty news")
             data = []
-        self.lock.release()
         self.closeDB()
         try:
             data = sorted(data, key=lambda anews: anews[1])
@@ -146,21 +139,19 @@ class wiredDB():
 
     def saveNews(self, news):
         if not self.openDB():
+            self.lock.release()
             return 0
-        self.lock.acquire()
         self.pointer.execute("INSERT INTO wiredNews VALUES (?,?,?);", [news[0], news[1], news[2]])
         self.conn.commit()
-        self.lock.release()
         self.closeDB()
         return 1
 
     def dropNews(self):
         if not self.openDB():
+            self.lock.release()
             return 0
-        self.lock.acquire()
         self.pointer.execute("DELETE FROM wiredNews")
         self.conn.commit()
-        self.lock.release()
         self.closeDB()
         return 1
 
@@ -170,11 +161,9 @@ class wiredDB():
             return 0
         sql = 'create table if not exists wiredIndex (name TEXT, type TEXT, size INTEGER, created FLOAT, \
               modified FLOAT, PRIMARY KEY(name));'
-        self.lock.acquire()
         self.pointer.execute(sql)
         self.pointer.execute('PRAGMA journal_mode=WAL;')
         self.conn.commit()
-        self.lock.release()
         return 1
 
     def closeIndex(self):
@@ -191,33 +180,26 @@ class wiredDB():
             self.pointer.execute("INSERT OR REPLACE INTO wiredIndex VALUES (?, ?, ?, ?, ?);",
                                  [str(aitem['name']), str(aitem['type']), str(aitem['size']), str(aitem['created']),
                                   str(aitem['modified'])])
-        self.lock.acquire()
         self.conn.commit()
-        self.lock.release()
         self.closeIndex()
         return 1
 
     def searchIndex(self, searchstring):
         self.openIndex()
         try:
-            self.lock.acquire()
             self.pointer.execute("SELECT * FROM wiredIndex WHERE name LIKE ?;", ['%' + str(searchstring) + '%'])
             result = self.pointer.fetchall()
         except:
             self.logger.error("Failed to process sqlite query for search term: %s", searchstring)
             self.closeIndex()
-            self.lock.release()
             return 0
-        self.lock.release()
         self.closeIndex()
         return result
 
     def pruneIndex(self, config, filelist):
         self.openIndex()
         self.pointer.execute("SELECT * FROM wiredIndex;")
-        self.lock.acquire()
         data = self.pointer.fetchall()
-        self.lock.release()
         lookup = {}
         for afile in filelist:
             lookup[afile['name']] = afile['type']
@@ -226,24 +208,19 @@ class wiredDB():
                 if lookup[aitem[0]] == aitem[1]:
                     continue  # all good
             # file vanished - remove from index
-            self.lock.acquire()
             self.pointer.execute("DELETE FROM wiredIndex WHERE name = ?", [aitem[0]])
             self.conn.commit()
-            self.lock.release()
-        self.lock.acquire()
         self.pointer.execute("VACUUM")
         self.conn.commit()
-        self.lock.release()
         self.closeIndex()
         return 1
 
     def getServerSize(self):
         self.openIndex()
-        self.lock.acquire()
         self.pointer.execute("SELECT SUM(size) FROM wiredIndex WHERE TYPE = 'file';")
         result = self.pointer.fetchone()
         self.conn.commit()
-        self.lock.release()
+        self.closeIndex()
         try:
             result = int(result[0])
         except TypeError:
@@ -253,12 +230,10 @@ class wiredDB():
 
     def getServerFiles(self):
         self.openIndex()
-        self.lock.acquire()
         self.pointer.execute("SELECT Count(*) FROM wiredIndex WHERE TYPE = 'file';")
         result = self.pointer.fetchone()
         self.conn.commit()
         self.closeIndex()
-        self.lock.release()
         try:
             result = int(result[0])
         except TypeError:
@@ -269,18 +244,16 @@ class wiredDB():
     ## Banned Users ##
     def openBans(self):
         if not self.openDB():
+            self.lock.release()
             return 0
         sql = 'create table if not exists bans (user TEXT, nick TEXT, ip TEXT, ends REAL);'
         try:
-            self.lock.acquire()
             self.pointer.execute(sql)
             self.pointer.execute('PRAGMA journal_mode=WAL;')
             self.conn.commit()
         except sqlite3.Error:
             self.logger.error("Failed to create ban table.")
-            self.lock.release()
             return 0
-        self.lock.release()
         return 1
 
     def closeBans(self):
@@ -291,32 +264,30 @@ class wiredDB():
         return 1
 
     def addBan(self, user, nick, ip, ends):
-        self.openBans()
+        if not self.openBans():
+            self.lock.release()
+            return 0
         try:
-            self.lock.acquire()
             self.pointer.execute("INSERT INTO bans VALUES (?,?,?,?);", [user, nick, ip, ends])
             self.conn.commit()
         except sqlite3.Error:
             self.logger.error("Failed to add ban to db")
             self.closeBans()
-            self.lock.release()
             return 0
-        self.lock.release()
         self.closeBans()
         return 1
 
     def checkBan(self, user, ip):
-        self.openBans()
+        if not self.openBans():
+            self.lock.release()
+            return 0
         try:
-            self.lock.acquire()
             self.pointer.execute("SELECT * FROM bans WHERE user = ? AND ip = ?;", [user, ip])
             data = self.pointer.fetchall()
         except sqlite3.Error:
             self.logger.error("Failed to query db bans")
-            self.lock.release()
             self.closeBans()
             return 0
-        self.lock.release()
         self.closeBans()
         if not data:
             return 0
@@ -332,18 +303,17 @@ class wiredDB():
         return 1
 
     def removeBan(self, user, nick, ip, ends):
-        self.openBans()
+        if not self.openBans():
+            self.lock.release()
+            return 0
         try:
-            self.lock.acquire()
             self.pointer.execute("DELETE FROM bans WHERE user = ? AND nick = ? AND ip = ? AND ends = ?;",
                                  [user, nick, ip, ends])
             self.conn.commit()
             self.pointer.execute("VACUUM")
         except sqlite3.Error:
-            self.lock.release()
             self.closeBans()
             self.logger.error("Failed to delete ban from DB")
             return 0
-        self.lock.release()
         self.closeBans()
         return 1
